backend/requirements.txt
# backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
pandas==2.1.3
reportlab==4.0.7